requirements.txt
flask==2.3.2
google-generativeai==0.4.0
flask-cors==4.0.0
